pyproject.toml
[project]
name = "vectorless-rag-service"
version = "0.1.0"
description = "Vectorless RAG microservice with PageIndex-style hierarchical retrieval."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Codex"}]
classifiers = [
  "Programming Language :: Python :: 3.11",
  "Typing :: Typed",
]

dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.8.2",
  "pydantic-settings==2.5.2",
  "python-multipart==0.0.9",
  "sqlalchemy==2.0.34",
  "psycopg2-binary==2.9.9",
  "alembic==1.13.2",
  "structlog==24.4.0",
  "opentelemetry-api==1.26.0",
  "opentelemetry-sdk==1.26.0",
  "opentelemetry-exporter-otlp==1.26.0",
  "opentelemetry-instrumentation-fastapi==0.47b0",
  "opentelemetry-instrumentation-requests==0.47b0",
  "prometheus-client==0.20.0",
  "PyPDF2==3.0.1",
  "rapidfuzz==3.9.7",
  "httpx==0.27.2",
  "boto3==1.35.10",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3",
  "pytest-asyncio==0.24.0",
  "ruff==0.6.9",
  "mypy==1.11.2",
  "types-requests==2.32.0.20240914",
]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
plugins = []

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"
